build(nix/devshell): Add `patch-nargo-backend` script
The `assert_binary_exists` function defined in
`tooling/backend_interface/src/lib.rs` downloads a pre-built noir backend
binary. This is a breaking change for developers working on NixOS, as this
binary can't be executed as is without manual patching.

This change adds a small script to the devshell that automates the patching
process until a proper solution is implemented - building the ACVM Barretenberg
backend via Nix.